Small Fixes

+++ main-mobile-app/app/src/main/java/feup/cmpv/feup/casadamusica/utils/Archive.java
@@ -391,6 +391,7 @@ public class Archive {
     }
 
     public static void removeVouchers(List<String> vouchersToRemove) {
+        if(vouchersToRemove == null) return;
         DBHelper db = new DBHelper(ApplicationContextRetriever.getContext());
         db.deleteVouchers(vouchersToRemove);
     }
